Fix parser for patterns w/multiple options (#1)
The pattern parser failed (returned invalid results) when trying to
parse layout patterns containing multiple conversion words that have
options. This was mainly due to greedy regex patterns for "group" and
"option"; and to incorrect post-match parsing.

This patch fixes the regex patterns and the post-match parsing. A unit
test was also added to test for this specific problem.